+++ test/resources/sets.clj
@@ -1,0 +1,4 @@
+(ns resources.sets)
+
+(defn killit [coll]
+  (not-any? #{"string1" "string2"} (map ffirst coll)))